+++ packages/apptive_grid_theme/test/apptive_grid_theme_test.dart
@@ -19,8 +19,8 @@ void main() {
     int columns = 1,
   }) {
     final themes = {
-      'Light': ApptiveGridTheme(brightness: Brightness.light).theme(),
-      'Dark': ApptiveGridTheme(brightness: Brightness.dark).theme(),
+      'Light': ApptiveGridTheme.create(brightness: Brightness.light),
+      'Dark': ApptiveGridTheme.create(brightness: Brightness.dark),
     };
     return goldenTest(
       description,