+++ packages/apptive_grid_form/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.12.4
+* Use Border of Theme for Input Textfields for Geolocation and References
+* Fix Hint Placement in User and (Multi)CrossReference Widget if no Item is selected
+
 ## 0.12.3
 * Thumbnails for Attachments
 * Attachments are now using their path when being added